更新了一个新烧饼的yaw

+++ New/SB_Up/User/AlgorithmLayer/handle_value.h
@@ -3,6 +3,6 @@
 #include "main.h"
 
 float remote_value(int16_t x, int16_t y);
-int16_t motor_value(int16_t get, int16_t set);
+float motor_value(int16_t get, int16_t set);
 
 #endif
